Create trait for update actions..
.. and make a blanket impl for it
